Refurbish of cross-sections

+++ ChiTech/ChiTimer/chi_timer_01_reset_get.cpp
@@ -50,7 +50,7 @@ std::string ChiTimer::GetLocalDateTimeString()
   std::time_t now = system_clock::to_time_t(system_clock::now());
 
   char s[30];
-  int end = std::strftime(s, 30, "%Y-%m-%d %H:%M:%S", std::localtime(&now));
+  size_t end = std::strftime(s, 30, "%Y-%m-%d %H:%M:%S", std::localtime(&now));
   s[29] = '\0';
   if (end < 30)  s[end]='\0';
   return s;
